Add graph as last template argument instead of first
If the query uses a template argument for the graph, supply it as the
last argument, not as the first one. This means that you can simply drop
these clauses from the query and the template will continue to work,
without having to renumber the previous argument 2 to 1, 3 to 2, etc.