Remove FROM from queries if there is no graph
If we don’t have a graph, we can’t substitute anything into FROM <%1$s>
patterns and the like in queries (the previous code would substitute a
stringified “null” in that case, which is highly unlikely to be the
correct thing). Instead, completely drop such clauses in that case, and
assume that operating on the default graph was what the user wanted.