Remove non-preload mode of runQuery()
When not in preload mode, the QueryExecution is never closed, which can
leak HTTP connections and eventually exhaust the connection pool. Since
RDF2Graph always calls this method in preload mode, we can simply get
rid of the other mode.

(Note that the QueryExecution is also not closed if an exception occurs,
but that will be fixed in a separate commit.)